Move doc and publication pages up
index.js should be avoided because it creates just one `index.js` bundle
with unrelated stuff.
